Was trying to add interpolation between rows, but it was already there :)

+++ Graffathon/Assets/Scripts/Rocket.cs
@@ -8,9 +8,9 @@ public class Rocket : Singleton<Rocket> {
 	public float speed;
 	public float bpm;
 	private bool playing;
-	private int row;
+	private float row;
 	float startTime;
-	private int startRow;
+	private float startRow;
 	private Dictionary<string, Track> tracks = new Dictionary<string, Track>();
 
 	// Use this for initialization
@@ -26,10 +26,10 @@ public class Rocket : Singleton<Rocket> {
 	void Update () {
 //		int currentRow;
 		if (playing)
-			row = ((int)((Time.time - startTime) * (bpm / 60f) + startRow));
+			row = ((Time.time - startTime) * (bpm / 60f) + startRow);
 //		else 
 //			currentRow = 0;
-		device.Update(row);
+		device.Update((int)row);
 		Debug.Log(row);
 	}
 
@@ -56,7 +56,39 @@ public class Rocket : Singleton<Rocket> {
 			tracks.Add(trackName, device.GetTrack(trackName));
 		}
 		return tracks[trackName].GetValue(row);
+//		if(!playing)
+//			return currentRow;
+//		float nextRow = tracks[trackName].GetValue(row);
+//		return currentRow + (nextRow - currentRow);
+//		return interpolate(currentRow, nextRow, row, tracks[trackName].Type);
 	}
+
+	/**
+     * Interpolates value for row between two TrackKey-objects.
+     *
+     * Interpolation type depends on the first key; only the value is
+     * used from the second key in the process.
+     *
+     * Row can naturally be fractional, so interpolation between
+     * values flowing in time is continuous.
+     */
+//	public static float interpolate(float first, float second, float row, Track.Key.Type type) {
+//		float t = (row - first) / (second - first);
+//			
+//		switch (type) {
+//		case Track.Key.Type.Step:
+//			return first;
+//			//case KeyType.LINEAR:
+//		case Track.Key.Type.Smooth:
+//			t = t * t * (3 - 2*t);
+//			break;
+//		case Track.Key.Type.Ramp:
+//			t = Mathf.Pow(t, 2.0);
+//			break;
+//		}
+//
+//		return first + (second - first) * t;
+//	}
 
 	void OnDestroy() {
 		device.Dispose();
